feat(v1-14): dialog & entry pages - decompose + tokenize + TDD redo (#1197)
