+++ apps/desktop/renderer/src/features/projects/CreateTemplateDialog.test.tsx
@@ -4,7 +4,10 @@ import userEvent from "@testing-library/user-event";
 import { CreateTemplateDialog } from "./CreateTemplateDialog";
 import { useTemplateStore } from "../../stores/templateStore";
 
-// Mock the template store
+// =============================================================================
+// Mock 设置
+// =============================================================================
+
 vi.mock("../../stores/templateStore", () => ({
   useTemplateStore: vi.fn(),
 }));
@@ -17,7 +20,6 @@ describe("CreateTemplateDialog", () => {
   beforeEach(() => {
     vi.clearAllMocks();
 
-    // Mock useTemplateStore to return our mock function
     (
       useTemplateStore as unknown as ReturnType<typeof vi.fn>
     ).mockImplementation(
@@ -34,10 +36,10 @@ describe("CreateTemplateDialog", () => {
   });
 
   // ===========================================================================
-  // 基础渲染测试
+  // 渲染 — 对话框结构与可见性
   // ===========================================================================
   describe("渲染", () => {
-    it("打开时应该渲染对话框", () => {
+    it("open 为 true 时渲染对话框表单", () => {
       render(
         <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
       );
@@ -45,7 +47,7 @@ describe("CreateTemplateDialog", () => {
       expect(screen.getByTestId("create-template-dialog")).toBeInTheDocument();
     });
 
-    it("关闭时不应该渲染对话框内容", () => {
+    it("open 为 false 时不渲染对话框内容", () => {
       render(
         <CreateTemplateDialog open={false} onOpenChange={mockOnOpenChange} />,
       );
@@ -55,15 +57,17 @@ describe("CreateTemplateDialog", () => {
       ).not.toBeInTheDocument();
     });
 
-    it("应该渲染模板名称输入框", () => {
+    it("显示名称输入框，带 autoFocus", () => {
       render(
         <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
       );
 
-      expect(screen.getByTestId("create-template-name")).toBeInTheDocument();
+      const nameInput = screen.getByTestId("create-template-name");
+      expect(nameInput).toBeInTheDocument();
+      expect(nameInput).toHaveFocus();
     });
 
-    it("应该渲染描述输入框", () => {
+    it("显示描述输入框", () => {
       render(
         <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
       );
@@ -73,20 +77,34 @@ describe("CreateTemplateDialog", () => {
       ).toBeInTheDocument();
     });
 
-    it("应该渲染创建按钮", () => {
+    it("显示创建和取消按钮", () => {
       render(
         <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
       );
 
       expect(screen.getByTestId("create-template-submit")).toBeInTheDocument();
+      expect(screen.getByText("Cancel")).toBeInTheDocument();
+    });
+
+    it("显示文件夹和文件添加区域", () => {
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      expect(
+        screen.getByPlaceholderText("e.g., chapters"),
+      ).toBeInTheDocument();
+      expect(
+        screen.getByPlaceholderText("e.g., outline.md"),
+      ).toBeInTheDocument();
     });
   });
 
   // ===========================================================================
-  // 表单提交测试
+  // 验证 — 名称必填校验
   // ===========================================================================
-  describe("表单提交", () => {
-    it("没有名称时不应该提交", async () => {
+  describe("验证", () => {
+    it("名称为空时提交不调用 createTemplate", async () => {
       const user = userEvent.setup();
 
       render(
@@ -98,7 +116,37 @@ describe("CreateTemplateDialog", () => {
       expect(mockCreateTemplate).not.toHaveBeenCalled();
     });
 
-    it("输入名称后应该能提交", async () => {
+    it("名称为空时提交显示验证错误消息", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      await user.click(screen.getByTestId("create-template-submit"));
+
+      expect(screen.getByText("Template name is required")).toBeInTheDocument();
+    });
+
+    it("仅空格的名称提交不调用 createTemplate", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      await user.type(screen.getByTestId("create-template-name"), "   ");
+      await user.click(screen.getByTestId("create-template-submit"));
+
+      expect(mockCreateTemplate).not.toHaveBeenCalled();
+    });
+  });
+
+  // ===========================================================================
+  // 表单提交 — 正常流程与回调
+  // ===========================================================================
+  describe("表单提交", () => {
+    it("输入名称后提交传递正确参数", async () => {
       const user = userEvent.setup();
 
       render(
@@ -124,7 +172,37 @@ describe("CreateTemplateDialog", () => {
       });
     });
 
-    it("提交成功后应该调用 onCreated", async () => {
+    it("名称和描述都填写时传递正确参数", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog
+          open={true}
+          onOpenChange={mockOnOpenChange}
+          onCreated={mockOnCreated}
+        />,
+      );
+
+      await user.type(
+        screen.getByTestId("create-template-name"),
+        "My Template",
+      );
+      await user.type(
+        screen.getByTestId("create-template-description"),
+        "A great template",
+      );
+      await user.click(screen.getByTestId("create-template-submit"));
+
+      await waitFor(() => {
+        expect(mockCreateTemplate).toHaveBeenCalledWith({
+          name: "My Template",
+          description: "A great template",
+          structure: { folders: [], files: [] },
+        });
+      });
+    });
+
+    it("提交成功后调用 onCreated 回调并传递模板 ID", async () => {
       const user = userEvent.setup();
 
       render(
@@ -146,7 +224,7 @@ describe("CreateTemplateDialog", () => {
       });
     });
 
-    it("提交成功后应该关闭对话框", async () => {
+    it("提交成功后自动关闭对话框", async () => {
       const user = userEvent.setup();
 
       render(
@@ -163,13 +241,8 @@ describe("CreateTemplateDialog", () => {
         expect(mockOnOpenChange).toHaveBeenCalledWith(false);
       });
     });
-  });
 
-  // ===========================================================================
-  // 文件夹/文件管理测试
-  // ===========================================================================
-  describe("文件夹管理", () => {
-    it("应该能添加文件夹", async () => {
+    it("添加文件夹和文件后提交包含完整 structure", async () => {
       const user = userEvent.setup();
 
       render(
@@ -180,66 +253,188 @@ describe("CreateTemplateDialog", () => {
         />,
       );
 
-      // Find folder input (placeholder: "e.g., chapters")
-      const folderInputs = screen.getAllByPlaceholderText("e.g., chapters");
-      await user.type(folderInputs[0], "my-folder");
-
-      // Click add button
+      // 添加文件夹
+      await user.type(
+        screen.getByPlaceholderText("e.g., chapters"),
+        "chapters",
+      );
       const addButtons = screen.getAllByText("Add");
       await user.click(addButtons[0]);
 
-      // Should show the folder in the list
+      // 添加文件
+      await user.type(
+        screen.getByPlaceholderText("e.g., outline.md"),
+        "readme.md",
+      );
+      await user.click(addButtons[1]);
+
+      // 提交
+      await user.type(
+        screen.getByTestId("create-template-name"),
+        "Full Template",
+      );
+      await user.click(screen.getByTestId("create-template-submit"));
+
+      await waitFor(() => {
+        expect(mockCreateTemplate).toHaveBeenCalledWith({
+          name: "Full Template",
+          description: undefined,
+          structure: {
+            folders: ["chapters"],
+            files: [{ path: "readme.md" }],
+          },
+        });
+      });
+    });
+  });
+
+  // ===========================================================================
+  // 文件夹管理 — 添加、移除、重复防护
+  // ===========================================================================
+  describe("文件夹管理", () => {
+    it("输入文件夹名并点击 Add 后显示在列表中", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      await user.type(
+        screen.getByPlaceholderText("e.g., chapters"),
+        "my-folder",
+      );
+      const addButtons = screen.getAllByText("Add");
+      await user.click(addButtons[0]);
+
       expect(screen.getByText("my-folder")).toBeInTheDocument();
     });
 
-    it("应该能移除文件夹", async () => {
+    it("按 Enter 键添加文件夹", async () => {
       const user = userEvent.setup();
 
       render(
         <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
       );
 
-      // Add a folder first
-      const folderInputs = screen.getAllByPlaceholderText("e.g., chapters");
-      await user.type(folderInputs[0], "my-folder");
+      const input = screen.getByPlaceholderText("e.g., chapters");
+      await user.type(input, "enter-folder{enter}");
+
+      expect(screen.getByText("enter-folder")).toBeInTheDocument();
+    });
+
+    it("添加后输入框自动清空", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      const input = screen.getByPlaceholderText("e.g., chapters");
+      await user.type(input, "my-folder");
       const addButtons = screen.getAllByText("Add");
       await user.click(addButtons[0]);
 
-      // Now remove it
-      const removeButton = screen.getByLabelText("Remove my-folder");
-      await user.click(removeButton);
-
-      // Should not be in the document anymore
-      expect(screen.queryByText("my-folder")).not.toBeInTheDocument();
+      expect(input).toHaveValue("");
     });
-  });
 
-  describe("文件管理", () => {
-    it("应该能添加文件", async () => {
+    it("点击移除按钮删除文件夹", async () => {
       const user = userEvent.setup();
 
       render(
         <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
       );
 
-      // Find file input (placeholder: "e.g., outline.md")
-      const fileInputs = screen.getAllByPlaceholderText("e.g., outline.md");
-      await user.type(fileInputs[0], "readme.md");
-
-      // Click add button (second one)
+      await user.type(
+        screen.getByPlaceholderText("e.g., chapters"),
+        "my-folder",
+      );
       const addButtons = screen.getAllByText("Add");
-      await user.click(addButtons[1]);
+      await user.click(addButtons[0]);
+      expect(screen.getByText("my-folder")).toBeInTheDocument();
 
-      // Should show the file in the list
-      expect(screen.getByText("readme.md")).toBeInTheDocument();
+      await user.click(screen.getByLabelText("Remove my-folder"));
+      expect(screen.queryByText("my-folder")).not.toBeInTheDocument();
+    });
+
+    it("重复文件夹名不会添加第二次", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      const input = screen.getByPlaceholderText("e.g., chapters");
+      const addButtons = screen.getAllByText("Add");
+
+      await user.type(input, "chapters");
+      await user.click(addButtons[0]);
+      await user.type(input, "chapters");
+      await user.click(addButtons[0]);
+
+      const items = screen.getAllByText("chapters");
+      expect(items).toHaveLength(1);
     });
   });
 
   // ===========================================================================
-  // 取消测试
+  // 文件管理 — 添加文件
+  // ===========================================================================
+  describe("文件管理", () => {
+    it("输入文件名并添加后显示在列表中", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      await user.type(
+        screen.getByPlaceholderText("e.g., outline.md"),
+        "readme.md",
+      );
+      const addButtons = screen.getAllByText("Add");
+      await user.click(addButtons[1]);
+
+      expect(screen.getByText("readme.md")).toBeInTheDocument();
+    });
+
+    it("按 Enter 键添加文件", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      const input = screen.getByPlaceholderText("e.g., outline.md");
+      await user.type(input, "notes.txt{enter}");
+
+      expect(screen.getByText("notes.txt")).toBeInTheDocument();
+    });
+
+    it("重复文件名不会添加第二次", async () => {
+      const user = userEvent.setup();
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      const input = screen.getByPlaceholderText("e.g., outline.md");
+      const addButtons = screen.getAllByText("Add");
+
+      await user.type(input, "readme.md");
+      await user.click(addButtons[1]);
+      await user.type(input, "readme.md");
+      await user.click(addButtons[1]);
+
+      const items = screen.getAllByText("readme.md");
+      expect(items).toHaveLength(1);
+    });
+  });
+
+  // ===========================================================================
+  // 取消 — 关闭对话框
   // ===========================================================================
   describe("取消", () => {
-    it("点击 Cancel 应该关闭对话框", async () => {
+    it("点击 Cancel 调用 onOpenChange(false)", async () => {
       const user = userEvent.setup();
 
       render(
@@ -253,10 +448,10 @@ describe("CreateTemplateDialog", () => {
   });
 
   // ===========================================================================
-  // 错误处理测试
+  // 错误处理 — 创建失败反馈
   // ===========================================================================
   describe("错误处理", () => {
-    it("创建失败时应该显示错误", async () => {
+    it("创建失败时显示错误消息", async () => {
       const user = userEvent.setup();
       mockCreateTemplate.mockRejectedValueOnce(new Error("Network error"));
 
@@ -272,6 +467,27 @@ describe("CreateTemplateDialog", () => {
 
       await waitFor(() => {
         expect(screen.getByText("Network error")).toBeInTheDocument();
+      });
+    });
+
+    it("非 Error 对象异常时显示通用错误消息", async () => {
+      const user = userEvent.setup();
+      mockCreateTemplate.mockRejectedValueOnce("string error");
+
+      render(
+        <CreateTemplateDialog open={true} onOpenChange={mockOnOpenChange} />,
+      );
+
+      await user.type(
+        screen.getByTestId("create-template-name"),
+        "My Template",
+      );
+      await user.click(screen.getByTestId("create-template-submit"));
+
+      await waitFor(() => {
+        expect(
+          screen.getByText("Failed to create template"),
+        ).toBeInTheDocument();
       });
     });
   });
